+++ src/bindings/3d_vector_bindings.h
@@ -1,0 +1,19 @@
+/*
+    src/bindings/3d_vector_bindings.h: Exposing ON_3dVector to Python
+
+    Copyright (c) 2026 Studio W Engineers
+
+    All rights reserved. Use of this source code is governed by a
+    BSD-style license that can be found in the LICENSE file.
+*/
+#pragma once
+
+// System includes
+
+// External includes
+
+// Project includes
+#include "bindings.h"
+
+
+void Vector3dBindings(nb::module_& m);