+++ docs/changelog.rst
@@ -8,6 +8,7 @@ Changelog
 Version 0.2.0 (T.B.A.)
 -----------------------------
 
+- Added bindings for ``ON_3dVector``.
 
 
 Version 0.1.1 (June 22, 2026)